Refactor: Consolidate plot_utils.py into visualization_utils.py
- Moved all functions and PlotStyleConfig from plot_utils.py into visualization_utils.py.
- Updated visualization_utils.py to use the local versions of these utilities.
- Removed redundant styling code in visualization_utils.py that is now handled by PlotStyleConfig and associated helpers.
- Refactored various plotting functions within visualization_utils.py to use the standardized helper functions for figure creation, styling, and saving.
- Deleted the original plot_utils.py file.
- Added a dummy CSV and updated test script path for basic execution testing.
